refactor(fourslash): split session client runner patches
